feat: add firefox extension with page action
Add manifest specifying page_action, content script and icons.
Add pageScript.js to be injected in the webpage to get the hash needed
for computing requests md5 and to use md5 function the webpage use.

The content script inject the pageScript.js and wait for a request from
popup script. On REFRESH_VALUES request, the script post a getData
request to retrieve city entities along with several details such as
squadSize, relics or cultural bonus and it emits a second request,
getRankingOverview to retrieve encounters and tournament scores.
On GET_VALUES it behaves the same except when it already has the data
stored from a previous call.

The popup script ask the content script about the base64 encoded text
reprensenting the city, the username, player_id and world/country to
either download the text in a file named after the user, world and
country, or to go to ElvenStats profile with player_id information, or
to upload the text directly to ElvenArchitect.
Currently, no production is set.

Signed-off-by: Olmagzar <olmagzar@gmail.com>